fix(a11y): server-render the landing with <main> + <h1>
Audit 35.2 A11Y-001/002 + LH-A11Y-HOME: `/` was a client component that emitted
an empty document on SSR (it returned null until the client-side silent refresh
resolved), so screen readers and crawlers saw no <main> / <h1>.

- page.tsx is now a server component: it reads the httpOnly `refresh_token`
  cookie and passes an `authHint` to the client. The anonymous landing is
  server-rendered (content in the initial HTML → crawlable + a11y; also improves
  LCP, addresses LH-PERF-HOME).
- home-content.tsx (new client component): the dual-state swap. The first client
  render uses `authHint` so it matches SSR (no hydration mismatch) and
  authenticated users don't flash the landing; once the silent refresh resolves
  the auth store is authoritative (a stale cookie falls back to the landing).
- landing-page.tsx: root `<div>` → `<main>` (the landmark); the `<h1>` already
  lives in landing/hero.tsx.

Verified: prettier + eslint + tsc clean on the changed files. SSR-safety, the
dual-state and hydration are exercised by the CI build + mocked/recette landing
suites; the landing VR output is unchanged (semantic-only div→main).
