+++ pwa/src/app/page.tsx
@@ -1,74 +1,20 @@
-"use client";
-
-import { Suspense, useEffect, useRef, useState } from "react";
-import { useAuthStore } from "@/store/auth-store";
-import { HydrationBoundary } from "@/components/hydration-boundary";
-import { TripPlanner } from "@/components/trip-planner";
-import { TripPlannerErrorBoundary } from "@/components/trip-planner-error-boundary";
-import { LandingPage } from "@/components/landing-page";
+import { cookies } from "next/headers";
+import { HomeContent } from "@/components/home-content";
 
 /**
- * Home page.
+ * Home page (dual-state: anonymous landing / authenticated dashboard).
  *
- * - Authenticated users see the TripPlanner.
- * - Unauthenticated users see the public LandingPage with early access form.
- *
- * Since `/` is a public route (no AuthGuard redirect), this component performs
- * its own silent auth check on mount to avoid flashing the landing page to
- * users who have a valid refresh token.
+ * Server component so the anonymous landing is server-rendered with its `<main>`
+ * and `<h1>` in the initial HTML (audit 35.2 A11Y-001/002 + LH-A11Y) — crawlable
+ * and accessible, instead of the previous client-only render that emitted an
+ * empty document. The presence of the httpOnly `refresh_token` cookie is passed
+ * as a hint so authenticated users render the dashboard directly (no landing
+ * flash) without a hydration mismatch; the client silent refresh then confirms.
  */
-function HomeContent() {
-  const isAuthenticated = useAuthStore((s) => s.isAuthenticated);
-  // When the user is already authenticated on mount we can render immediately;
-  // otherwise we must await the silent refresh before deciding what to render.
-  const [refreshDone, setRefreshDone] = useState(false);
-  const checkStarted = useRef(false);
+export default async function Page() {
+  const cookieStore = await cookies();
+  // Cookie name mirrors api `AuthCookies::REFRESH_TOKEN`.
+  const authHint = cookieStore.get("refresh_token")?.value ? "authed" : "anon";
 
-  useEffect(() => {
-    if (checkStarted.current) return;
-    checkStarted.current = true;
-
-    if (isAuthenticated) return;
-
-    const check = async () => {
-      try {
-        await useAuthStore.getState().silentRefresh();
-      } finally {
-        setRefreshDone(true);
-      }
-    };
-
-    void check();
-  }, [isAuthenticated]);
-
-  // Render nothing while checking auth to avoid flash
-  if (!isAuthenticated && !refreshDone) {
-    return null;
-  }
-
-  if (!isAuthenticated) {
-    return (
-      <Suspense fallback={null}>
-        <LandingPage />
-      </Suspense>
-    );
-  }
-
-  return (
-    <HydrationBoundary>
-      <TripPlannerErrorBoundary>
-        <Suspense fallback={null}>
-          <TripPlanner />
-        </Suspense>
-      </TripPlannerErrorBoundary>
-    </HydrationBoundary>
-  );
-}
-
-export default function Page() {
-  return (
-    <Suspense fallback={null}>
-      <HomeContent />
-    </Suspense>
-  );
+  return <HomeContent authHint={authHint} />;
 }
